misc. cleanup (#120)
* Resolve conversion from float warnings

* Mark player snail with Player component
- in case we want the NPC to also be tagged as a Snail later on

* Add NoCollide component tag to ignore equipped collectibles

+++ src/dialogue.cpp
@@ -50,7 +50,7 @@ void DialogueSystem::displayCurrent(float elapsed_ms)
 
 	const auto FANTASQUE_SANS_MONO_REGULAR = Font::load(FANTASQUE_SANS_MONO_REGULAR_PATH);
 
-	int chars_to_display = dialogue_displayed_ms / MS_PER_CHAR;
+	int chars_to_display = static_cast<int>(dialogue_displayed_ms / MS_PER_CHAR);
 	int chars_displayed = 0;
 	for (int i = 0; i < dialogueQueue.front().size(); i++)
 	{
